[feature/only-OP-can-view] Force field functions_posting.php
Applied the change to functions_posting.
It now prevents the user from viewing the topic link if he should not be able
to see it.
Also it checks if the user still has permissions to load a draft just like it
 is made with f_read permission.

For topic_review(), It applies the standard adding the extra data for query to
 the DB in case the user does not have the permission to read all threads in
 the forum so that it can be confirmed that the user really can or cannot view
 those posts

For user_notification(), The rule applies, included the check for the user_id
 and the f_read_other everywhere the f_read exists and changed a bit the way
 the verifications work so that it becomes a bit more efficient and to be
 easier to execute the tests needed.

PHPBB3-10623