Content_visibility; get_forums_visibility_sql() inject proposal
This allows making much less event points as nearly all situations where these methods are used are on SQL queries that require my action.
If rejected, nearly all places where this is used will require an event placement.
NOTE: I'm OK with the other situation if this one is rejected, just check which one is the best option.